[feat] 신고 버튼 공통 컴포넌트화
